Fix Porsche Vehicle implementation
- Make Odomoter, Climater, ChargeState, FinishTime APIs actually usable, they weren't exposed before
- Combine Provider and ProviderEmobility into one, as they complement each other instead of being different APIs for the same purpose
- Fix SoC not working since 27th October. The Emobility API only works if the Status API is called before. Otherwise the Emobility API returns HTTP 502
