+++ vehicle/porsche.go
@@ -11,7 +11,11 @@ import (
 
 type porscheProvider interface {
 	api.Battery
+	api.ChargeState
 	api.VehicleRange
+	api.VehicleClimater
+	api.VehicleFinishTimer
+	api.VehicleOdometer
 }
 
 // Porsche is an api.Vehicle implementation for Porsche cars
@@ -52,12 +56,7 @@ func NewPorscheFromConfig(other map[string]interface{}) (api.Vehicle, error) {
 		return nil, err
 	}
 
-	var provider porscheProvider
-	if vehicle.EmobilityVehicle {
-		provider = porsche.NewEMobilityProvider(log, identity, accessTokens.EmobilityToken, vehicle.VIN, cc.Cache)
-	} else {
-		provider = porsche.NewProvider(log, identity, accessTokens.Token, vehicle.VIN, cc.Cache)
-	}
+	provider := porsche.NewProvider(log, identity, accessTokens, vehicle.VIN, cc.Cache)
 
 	v := &Porsche{
 		embed:           &cc.embed,
